% Generated by roxygen2 (4.0.2): do not edit by hand
\name{app_shot}
\alias{app_shot}
\title{Take a screenshot of a Shiny app}
\usage{
app_shot(app, file, ..., port = 9000)
}
\arguments{
\item{app}{A Shiny app object, or a string naming an app directory.}

\item{file}{Name of output file. Should end with \code{.png}.}

\item{...}{Other arguments to pass on to \code{\link{web_shot}}.}

\item{port}{Port that Shiny will listen on.}
}
\description{
Take a screenshot of a Shiny app
}
\examples{
\donttest{
appdir <- system.file("examples", "01_hello", package="shiny")
app_shot(appdir, "01_hello.png")
}
}

