% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appshot.R
\name{appshot}
\alias{appshot}
\alias{appshot.shiny.appobj}
\alias{appshot.character}
\title{Take a screenshot of a Shiny app}
\usage{
appshot(app, file = "webshot.png", ..., port = getOption("shiny.port"),
  envvars = NULL)

\method{appshot}{shiny.appobj}(app, file = "webshot.png", ...,
  port = getOption("shiny.port"), envvars = NULL, env = NULL,
  packages = NULL, save_file = tempfile(fileext = ".RData"),
  env_file = tempfile(fileext = ".RData"))

\method{appshot}{character}(app, file = "webshot.png", ...,
  port = getOption("shiny.port"), envvars = NULL)
}
\arguments{
\item{app}{A Shiny app object, or a string naming an app directory.}

\item{file}{A vector of names of output files. Should end with \code{.png},
\code{.pdf}, or \code{.jpeg}. If several screenshots have to be taken and
only one filename is provided, then the function appends the index number
of the screenshot to the file name.}

\item{...}{Other arguments to pass on to \code{\link{webshot}}.}

\item{port}{Port that Shiny will listen on.}

\item{envvars}{A named character vector or named list of environment
variables and values to set for the Shiny app's R process. These will be
unset after the process exits. This can be used to pass configuration
information to a Shiny app.}

\item{env}{Environment to export all R objects for the Shiny app object.}

\item{packages}{Character vector of R packages required for the Shiny app execution.  Defaults to use the currently attached packages.}

\item{save_file}{File used to temporarily save the Shiny app object.}

\item{env_file}{File used to temporarily save the Shiny app object server context to execute the app.}
}
\description{
Take a screenshot of a Shiny app
}
\examples{
if (interactive()) {
  appdir <- system.file("examples", "01_hello", package="shiny")

  # With a Shiny directory
  appshot(appdir, "01_hello.png")

  # With a Shiny App object
  shinyapp <- shiny::shinyAppDir(appdir)
  appshot(shinyapp, "01_hello_app.png")
}

}
