% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_phantomjs}
\alias{install_phantomjs}
\title{Install PhantomJS}
\usage{
install_phantomjs(version = "2.1.1")
}
\arguments{
\item{version}{The version number of PhantomJS.}
}
\value{
\code{NULL} (the executable is written to a system directory).
}
\description{
Download the zip package, unzip it, and copy the executable to a system
directory in which \pkg{webshot} can look for the PhantomJS executable.
Currently this function only works for Windows. Mac OS X users are
recommended to install PhantomJS via Homebrew. If you download the package
from the PhantomJS website instead, please make sure the executable can be
found via the \code{PATH} variable.
}

