% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_phantomjs_installed}
\alias{is_phantomjs_installed}
\title{Determine if PhantomJS is Installed}
\usage{
is_phantomjs_installed()
}
\value{
\code{TRUE} if the PhantomJS is installed. Otherwise, \code{FALSE}
if PhantomJS is not installed.
}
\description{
Verifies that a version of PhantomJS is installed and available for use
on the user's computer.
}
