% Generated by roxygen2 (4.0.2): do not edit by hand
\name{shrink}
\alias{shrink}
\title{Shrink file size of a PNG}
\usage{
shrink(filename)
}
\arguments{
\item{filename}{Name of image to shrink. Must be a PNG file.}
}
\description{
This does not change size of the image in pixels, nor does it affect
appearance -- it is lossless compression. This requires the program
\code{optipng} to be installed.
}
\examples{
\donttest{
# Can be chained with webshot() or appshot()
webshot("http://www.google.com/", "google-small.png") \%>\%
 resize("75\%") \%>\%
 shrink()
}
}

