% Generated by roxygen2 (4.0.2): do not edit by hand
\name{web_shot}
\alias{web_shot}
\title{Take a screenshot of a URL}
\usage{
web_shot(url = NULL, file = "webshot.png", vwidth = 920, vheight = 600,
  cliprect = NULL, selector = NULL, port = NULL)
}
\arguments{
\item{url}{A URL to visit.}

\item{file}{Name of output file. Should end with \code{.png}.}

\item{vwidth}{Viewport width. This is the width of the browser "window".}

\item{vheight}{Viewport height This is the height of the browser "window".}

\item{cliprect}{Clipping rectangle. If unspecified, the clipping rectangle
matches the viewport size. Otherwise, it should be a four-element numeric
vector specifying the top, left, width, and height. This option is not
compatible with \code{selector}.}

\item{selector}{A CSS selector specifying a DOM element to set the clipping
  rectangle to. A screenshot of just this DOM element will be taken. If the
  selector has more than one match, only the first one will be used. This
  option is not compatible with \code{cliprect}.}
}
\description{
Take a screenshot of a URL
}
\examples{
web_shot("http://www.rstudio.com/", "rstudio-header.png", selector = "#header")
web_shot("https://github.com/rstudio/shiny/", "shiny-stats.png",
         selector = "ul.numbers-summary")
}

