% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webshot.R
\name{webshot}
\alias{webshot}
\title{Take a screenshot of a URL}
\usage{
webshot(url = NULL, file = "webshot.png", vwidth = 992, vheight = 744,
  cliprect = NULL, selector = NULL, expand = NULL, delay = 0.2,
  eval = NULL)
}
\arguments{
\item{url}{A URL to visit.}

\item{file}{Name of output file. Should end with \code{.png}, \code{.pdf}, or
\code{.jpeg}.}

\item{vwidth}{Viewport width. This is the width of the browser "window".}

\item{vheight}{Viewport height This is the height of the browser "window".}

\item{cliprect}{Clipping rectangle. If \code{cliprect} and \code{selector}
are both unspecified, the clipping rectangle will contain the entire page.
This can be the string \code{"viewport"}, in which case the clipping
rectangle matches the viewport size, or it can be a four-element numeric
vector specifying the top, left, width, and height. This option is not
compatible with \code{selector}.}

\item{selector}{One or more CSS selectors specifying a DOM element to set the
clipping rectangle to. The screenshot will contain these DOM elements. For
a given selector, if it has more than one match, only the first one will be
used. This option is not compatible with \code{cliprect}.}

\item{expand}{A numeric vector specifying how many pixels to expand the
clipping rectangle by. If one number, the rectangle will be expanded by
that many pixels on all sides. If four numbers, they specify the top,
right, bottom, and left, in that order.}

\item{delay}{Time to wait before taking screenshot, in seconds. Sometimes a
longer delay is needed for all assets to display properly.}

\item{eval}{An optional string with JavaScript code which will be evaluated
after opening the page and waiting for \code{delay}, but before calculating
the clipping region and taking the screenshot. See the Casper API
(\url{http://docs.casperjs.org/en/latest/modules/casper.html}) for more
information about what commands can be used to control the web page. NOTE:
This is experimental and likely to change!}
}
\description{
Take a screenshot of a URL
}
\examples{
if (interactive()) webshot("https://github.com/rstudio/shiny")

# See more examples in the package vignette
vignette("intro", package = "webshot")
}
\seealso{
\code{\link{appshot}} for taking screenshots of Shiny applications.
}

