% Generated by roxygen2 (4.0.2): do not edit by hand
\name{appshot}
\alias{appshot}
\title{Take a screenshot of a Shiny app}
\usage{
appshot(app, file = "webshot.png", ..., port = 9000)
}
\arguments{
\item{app}{A Shiny app object, or a string naming an app directory.}

\item{file}{Name of output file. Should end with \code{.png}, \code{.pdf}, or
\code{.jpeg}.}

\item{...}{Other arguments to pass on to \code{\link{webshot}}.}

\item{port}{Port that Shiny will listen on.}
}
\description{
Take a screenshot of a Shiny app
}
\examples{
\donttest{
appdir <- system.file("examples", "01_hello", package="shiny")
appshot(appdir, "01_hello.png")
}
}

